requirements.txt
langchain-community
yandex-cloud-ml-sdk
faiss-cpu
yandexcloud
pypdf
python-dotenv
pytest
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.9.0
tqdm
python-pptx-1.0.2